pyproject.toml
[project]
name = "seo-professor"
version = "0.1.0"
description = "AI-powered SEO audit suite"
requires-python = ">=3.11"
dependencies = [
    "supabase>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "weasyprint>=62",
    "lxml>=5.0",
    "jinja2>=3.1",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    # Trigger.dev Python SDK — install: pip install trigger.dev
    # (PyPI package name: check https://trigger.dev/docs/sdk/python)
    "trigger-dev>=3.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.22",
]

[project.scripts]
seo-audit = "src.audit:main"
seo-web   = "src.web:start"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
